# docs/swagger.yaml
definitions:
  models.Address:
    properties:
      addressID:
        type: integer
      addressLine1:
        type: string
      addressLine2:
        type: string
      city:
        type: string
      state:
        type: string
      userID:
        type: integer
      zipCode:
        type: string
    type: object
...
  models.User:
    properties:
      address:
        items:
          $ref: '#/definitions/models.Address'
        type: array
      contactNumber:
        type: string
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      userID:
        type: integer
      verifiedAccount:
        type: boolean
    type: object
...
  services.AddProductWeightRequest:
    properties:
      price:
        type: number
      stock_availability:
        type: integer
      weight:
        type: integer
    type: object
  services.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  services.SuccessResponse:
    properties:
      message:
        type: string
      product_id:
        type: integer
    type: object
  services.UpdateProductWeightRequest:
    properties:
      price:
        type: number
      stock_availability:
        type: integer
      weight:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /productweight/{productID}/weights:
    post:
      parameters:
      - description: Product ID
        in: path
        name: productID
        required: true
        type: string
      - description: Product weight details
        in: body
        name: weight
        required: true
        schema:
          $ref: '#/definitions/services.AddProductWeightRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Weight added successfully
          schema:
            $ref: '#/definitions/services.SuccessResponse'
        "400":
          description: Invalid request body or product ID
          schema:
            $ref: '#/definitions/services.ErrorResponse'
        "500":
          description: Failed to add weight
          schema:
            $ref: '#/definitions/services.ErrorResponse'
      summary: Add a new weight variant for a product
      tags:
      - Product Weights
  /productweight/{productID}/weights/{weightID}:
    put:
      parameters:
      - description: Product ID
        in: path
        name: productID
        required: true
        type: string
      - description: Weight ID
        in: path
        name: weightID
        required: true
        type: string
      - description: Product weight details
        in: body
        name: weight
        required: true
        schema:
          $ref: '#/definitions/services.UpdateProductWeightRequest'
      responses:
        "200":
          description: Weight updated successfully
        "400":
          description: Invalid request body or product/weight ID
        "500":
          description: Failed to update weight
      summary: Update an existing weight variant for a product
      tags:
      - Product Weights
  /users:
    post:
      consumes:
      - application/json
      parameters:
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: User created successfully
          schema:
            type: string
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/services.ErrorResponse'
        "409":
          description: Contact number already exists
          schema:
            $ref: '#/definitions/services.ErrorResponse'
        "500":
          description: Failed to create user
          schema:
            $ref: '#/definitions/services.ErrorResponse'
      summary: Create a new user
      tags:
      - Users
  /users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: User deleted successfully
          schema:
            type: string
        "500":
          description: Failed to delete user
          schema:
            $ref: '#/definitions/services.ErrorResponse'
      summary: Delete a user
      tags:
      - Users
    get:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User retrieved successfully
          schema:
            $ref: '#/definitions/models.User'
        "500":
          description: Failed to retrieve user
          schema:
            $ref: '#/definitions/services.ErrorResponse'
      summary: Retrieve a user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: User updated successfully
          schema:
            type: string
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/services.ErrorResponse'
        "500":
          description: Failed to update user
          schema:
            $ref: '#/definitions/services.ErrorResponse'
      summary: Update a user
      tags:
      - Users
swagger: "2.0"